+++ app/request-time-off/page.tsx
@@ -1,0 +1,5 @@
+import RequestTimeOffSheet from "../sheets/request-time-off/RequestTimeOffSheet";
+
+export default function RequestTimeOffPage() {
+  return <RequestTimeOffSheet />;
+}