+++ app/sheets/registry.ts
@@ -6,6 +6,7 @@ import { hybridWeeklyDefinition } from "./hybrid-weekly/definition";
 import { interiorCleaningDefinition } from "./interior-cleaning/definition";
 import { meterReadingsDefinition } from "./meter-readings/definition";
 import { monthlyCleaningDefinition } from "./monthly-cleaning/definition";
+import { requestTimeOffDefinition } from "./request-time-off/definition";
 import {
   fareboxSheetSchema,
   fuelSheetSchema,
@@ -135,6 +136,7 @@ export const SHEET_DEFINITIONS = [
   hybridDailyDefinition,
   hybridWeeklyDefinition,
   monthlyCleaningDefinition,
+  requestTimeOffDefinition,
 ] as const;
 
 type RegisteredSheetDefinition = (typeof SHEET_DEFINITIONS)[number];